backend/recommandations_scraper/requirements.txt
python-dotenv
selenium
webdriver-manager
beautifulsoup4
google-generativeai
requests
google-cloud-aiplatform
selenium-stealth